% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMatrix.R
\docType{methods}
\name{apply,dMatrix-method}
\alias{apply,dMatrix-method}
\title{Apply function for rDMatrix or cDMatrix objects.}
\usage{
\S4method{apply}{dMatrix}(X, MARGIN, FUN, chunkSize = 1000, verbose = TRUE,
  ...)
}
\arguments{
\item{X}{Either an rDMatrix or a cDMatrix object}

\item{MARGIN}{Use 1 to obtain row summaries or 2 to obtain column summaries}

\item{chunkSize}{The number of columns or rows that are processed at a time
(see Details).}
}
\value{
Returns a matrix or a list with results from FUN.
}
\description{
This function brings chunks of data (of size chunkSize) from the distributed
array into RAM as matrix objects and calls the apply method of the base
package to obtain the summaries for the chunk. Results from all the chunks
are collected and returned.
}

