% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{as.BGData}
\alias{as.BGData}
\title{Convert Other Objects to BGData Objects.}
\usage{
as.BGData(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments (see method).}
}
\value{
A BGData object.
}
\description{
Convert Other Objects to BGData Objects.
}
\seealso{
\code{\link[=as.BGData.BEDMatrix]{as.BGData.BEDMatrix()}} to convert \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix} objects
to \linkS4class{BGData}.
}
