% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BGData}
\alias{BGData}
\alias{BGData-package}
\title{BGData: Memory Mapped Matrices and Data-Structures for Genomic Data for R}
\description{
More info in the wiki: \link{https://github.com/QuantGen/BGData/wiki}
}

