% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cDMatrix.R
\docType{class}
\name{cDMatrix-class}
\alias{cDMatrix}
\alias{cDMatrix-class}
\title{An S4 class to represent a column-distributed \code{mmMatrix}.}
\description{
\code{cDMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

