% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mmMatrix.R
\name{chunks}
\alias{chunks}
\title{Provides information about how data is distributed into binary files.}
\usage{
chunks(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{rmmMatrix}} or a
\code{\linkS4class{cmmMatrix}} object}
}
\value{
A matrix with information per chunk in rows.
}
\description{
Row-distributed (\code{\linkS4class{rmmMatrix}}) and column-distributed
matrices (\code{\linkS4class{cmmMatrix}}) have the content of the array mapped
to possibly multiple binary files. Each chunk is an \code{ff_matrix} object.
\code{chunks} gives, for each chunk, the row or column indexes at which each
chunk start and ends.
}

