% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmmMatrix.R
\docType{class}
\name{cmmMatrix-class}
\alias{cmmMatrix}
\alias{cmmMatrix-class}
\title{An S4 class to represent a column-distributed \code{mmMatrix}.}
\description{
\code{cmmMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

