% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cDMatrix.R
\name{colindexes}
\alias{colindexes}
\title{Finds the position of a set of columns in a cDMatrix object.}
\usage{
colindexes(x, columns)
}
\description{
Finds the position of a set of columns in a cDMatrix object.
}

