% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{colMeans,LinkedMatrix-method}
\alias{colMeans,LinkedMatrix-method}
\title{Form column means.}
\usage{
\S4method{colMeans}{LinkedMatrix}(x, na.rm = TRUE, chunkSize = 1000, ...)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric,
    complex, integer or logical values, or a numeric data frame.}

\item{na.rm}{logical.  Should missing values (including \code{NaN})
    be omitted from the calculations?}

\item{chunkSize}{The number of columns that are processed at a time.}

\item{...}{Optional arguments to \code{mean}.}
}
\description{
Form column means.
}

