% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genData.R
\docType{class}
\name{genData-class}
\alias{genData}
\alias{genData-class}
\title{An S4 class to represent GWAS data.}
\description{
An S4 class to represent GWAS data.
}
\section{Slots}{

\describe{
\item{\code{pheno}}{A \code{\link{data.frame}} that contains phenotypes.}

\item{\code{map}}{A \code{\link{data.frame}} that contains a genetic map.}

\item{\code{geno}}{A \code{geno} object (\code{dMatrix}, \code{ff_matrix}, or
\code{\link{matrix}}) that contains genotypes.}
}}

