% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\docType{class}
\name{geno-class}
\alias{geno-class}
\title{An abstract S4 class union of matrix-like types.}
\description{
\linkS4class{geno} is a class union of several matrix-like types, many of them
suitable for very large datasets. Currently supported are
\link[LinkedMatrix:LinkedMatrix-class]{LinkedMatrix::LinkedMatrix}, \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix},
\link[bigmemory:big.matrix-class]{bigmemory::big.matrix}, \code{ff_matrix}, and \code{matrix}.
}
