% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getG.symDMatrix}
\alias{getG.symDMatrix}
\title{Computes a genomic relationship matrix G=xx' without ever loading G in RAM by
creating a \code{\link[=symDMatrix-class]{symDMatrix}}.}
\usage{
getG.symDMatrix(X, i = seq_len(nrow(X)), j = seq_len(ncol(X)),
  centerCol = TRUE, centers = NULL, scaleCol = TRUE, scales = NULL,
  scaleG = TRUE, folder = randomString(), vmode = "double",
  saveRData = TRUE, bufferSize = NULL, nBuffers = NULL, nTasks = nCores,
  nCores = 1, verbose = TRUE)
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
xx'.}

\item{centers}{Precomputed centers.}

\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
xx'.}

\item{scales}{Precomputed scales.}

\item{scaleG}{TRUE/FALSE whether xx' must be scaled.}

\item{folder}{Folder in which to save the
\code{\link[=symDMatrix-class]{symDMatrix}}.}

\item{vmode}{vmode of \code{ff} objects.}

\item{saveRData}{Whether to save an RData file to easily reload
\code{\link[=symDMatrix-class]{symDMatrix}}}

\item{bufferSize}{The number of columns of \code{x} that are brought into
RAM for processing. Overwrites \code{nBuffers}. If both parameters are
\code{NULL}, all elements of \code{j} are used.}

\item{nBuffers}{The number of partitions of the columns of \code{x} that are
brought into RAM for processing. Is overwritten by \code{bufferSize}. If
both parameters are \code{NULL}, all elements of \code{j} are used.}

\item{nTasks}{The number of chunks that each chunk is split into for
processing in parallel.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}

\item{verbose}{If TRUE more messages are printed.}
}
\value{
A positive semi-definite symmetric numeric matrix.
}
\description{
Offers options for centering and scaling the columns of x before computing
xx'.
}

