% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMatrix.R
\name{GWAS}
\alias{GWAS}
\title{Conducts an association study (GWAS) using a genData object.}
\usage{
GWAS(formula, data, method, plot = FALSE, verbose = FALSE,
  min.pValue = 1e-10, chunkSize = 10, ...)
}
\arguments{
\item{formula}{A formula (e.g. weight~sex+age) with the response on the
left-hand side and predictors (all the covariates except the markers) on
the right-hand side. The variables included in the formula must be in the
@pheno object of the genData.}

\item{data}{A genData object.}

\item{method}{The regression method to be used. Currently, the following
methods are implemented: lm, lm.fit, lsfit, glm and lmer.}

\item{plot}{If TRUE a Manhattan plot is produced and filled with points as
the association tests are run.}

\item{verbose}{If TRUE more messages are printed.}

\item{min.pValue}{Numeric, the minimum p-value expected, used to determine
the limits of the vertical axis of the Manhattan plot.}

\item{chunkSize}{Represents the number of columns of @geno that are brought
into RAM for processing (10 by default).}
}
\value{
Returns a matrix with estimates, SE, p-value, etc.
}
\description{
This function conducts an association test using the formula provided by the
user (see formula above) plus one column of @geno, one column at a time. The
data from the association tests is obtained from a genData object.
}

