% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmMatrix.R
\name{index}
\alias{index}
\title{Finds the position of a set of rows or columns in an rmmMatrix or cmmMatrix object.}
\usage{
index(x)
}
\description{
Finds the position of a set of rows or columns in an rmmMatrix or cmmMatrix object.
}

