% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\docType{methods}
\name{initialize,BGData-method}
\alias{initialize,BGData-method}
\title{Creates a new \code{\link[=BGData-class]{BGData}} instance.}
\usage{
\S4method{initialize}{BGData}(.Object, geno, pheno, map)
}
\arguments{
\item{.Object}{The \code{\link[=BGData-class]{BGData}} instance to be
initialized.}

\item{geno}{A \code{geno} object that contains genotypes.}

\item{pheno}{A \code{\link{data.frame}} that contains phenotypes. A stub that
only contains an \code{IID} column populated with the rownames of 
\code{@geno} will be generated if missing.}

\item{map}{A \code{\link{data.frame}} that contains a genetic map. A stub 
that only contains a \code{mrk} column populated with the colnames of 
\code{@geno} will be generated if missing.}
}
\description{
Creates a new \code{\link[=BGData-class]{BGData}} instance.
}

