% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BGData.R
\name{loadBGData}
\alias{loadBGData}
\title{Loads a BGData object using the name of the folder where the metadata and
data are stored.}
\usage{
loadBGData(path, envir = .GlobalEnv)
}
\arguments{
\item{path}{The name of the folder where the data and meta data are stored.}

\item{envir}{The name of the environment where the object is returned.}
}
\description{
Loads a BGData object using the name of the folder where the metadata and
data are stored.
}
\seealso{
\code{\link{load2}} and \code{\link{readPED}}
}

