% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{orderedMerge}
\alias{orderedMerge}
\title{Merge Two Data Frames Keeping the Order of the First}
\usage{
orderedMerge(x, y, by = c(1L, 2L))
}
\arguments{
\item{x, }{y Data frames}

\item{by}{Specifications of the columns used for merging. Defaults to the
first two columns of the data frame, which traditionally has the family ID
and the individual ID.}
}
\value{
Merged data frame
}
\description{
This is a simplified version of \code{\link[base:merge]{base::merge()}} useful for merging
additional data into a \linkS4class{BGData} object while keeping the order of the
data in the \linkS4class{BGData} object.
}
