% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMatrix.R
\docType{class}
\name{rDMatrix-class}
\alias{rDMatrix}
\alias{rDMatrix-class}
\title{An S4 class to represent a row-distributed \code{dMatrix}}
\description{
\code{rDMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

