% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{readPED.big.matrix}
\alias{readPED.big.matrix}
\title{Creates a \code{\link[=BGData-class]{BGData}} object from a raw PED file
(generated with \code{--recodeA} in PLINK) or a PED-like file and stores the
genotypes in a \code{\link[=big.matrix-class]{big.matrix}}.}
\usage{
readPED.big.matrix(fileIn, header, dataType, n = NULL, p = NULL, sep = "",
  na.strings = "NA", nColSkip = 6, idCol = c(1, 2),
  folderOut = paste("BGData_", sub("\\\\.[[:alnum:]]+$", "",
  basename(fileIn)), sep = ""), verbose = FALSE)
}
\arguments{
\item{fileIn}{The path to the plaintext file.}

\item{header}{If TRUE, the file contains a header.}

\item{dataType}{The coding of genotypes. Use \code{character()} for A/C/G/T
or \code{integer()} for numeric coding.}

\item{n}{The number of individuals.}

\item{p}{The number of markers.}

\item{sep}{The field separator character. Values on each line of the file
are separated by this character. If \code{sep = ""} (the default for
\code{readPED.big.matrix}) the separator is "white space", that is one or
more spaces, tabs, newlines or carriage returns.}

\item{na.strings}{The character string used in the plaintext file to denote
missing value.}

\item{nColSkip}{The number of columns to be skipped to reach the genotype
information in the file.}

\item{idCol}{The index of the ID column. If more than one index is given,
both columns will be concatenated with "_".}

\item{folderOut}{The path to the folder where to save the binary files.}

\item{verbose}{If TRUE, progress updates will be posted.}
}
\value{
Returns a \code{\link[=BGData-class]{BGData}} object.
}
\description{
\code{readPED.matrix} assumes that the plaintext file (\code{fileIn})
contains records of individuals in rows, and phenotypes, covariates and
markers in columns. The columns included in columns \code{seq_len(nColSkip)} are
used to populate the slot \code{@pheno} of a
\code{\link[=BGData-class]{BGData}} object, and the remaining columns are
used to fill the slot \code{@geno}. If the first row contains a header
(\code{header=TRUE}), data in this row is used to determine variables names
for \code{@pheno} and marker names for \code{@map} and \code{@geno}.
}
\details{
Genotypes are stored in a filebacked
\code{\link[=big.matrix-class]{big.matrix}} object. \code{readPED.big.matrix}
creates a folder (see \code{folderOut}) that contains the binary flat file
(named \code{BGData.bin}), a descriptor file (named \code{BGData.desc}), and
an external representation of the \code{\link[=BGData-class]{BGData}} object
in \code{BGData.RData}. A \code{\link[=BGData-class]{BGData}} object can be
reloaded using \code{load.BGData} (the regular \code{load} function will not
work unless the \code{\link[=big.matrix-class]{big.matrix}} instance is
manually attached using \code{\link[bigmemory]{attach.big.matrix}}).
}
\seealso{
\code{\link[=BGData-class]{BGData}}
}

