% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmmMatrix.R
\docType{class}
\name{rmmMatrix-class}
\alias{rmmMatrix}
\alias{rmmMatrix-class}
\title{An S4 class to represent a row-distributed \code{mmMatrix}}
\description{
\code{rmmMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

