% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{simPED}
\alias{simPED}
\title{Generate and store a simulated plaintext raw PED file (see \code{--recodeA}
in PLINK) or PED-like file for testing purposes.}
\usage{
simPED(filename, n, p, genoChars = 0:2, na.string = NA, propNA = 0.02,
  returnGenos = FALSE)
}
\description{
Generate and store a simulated plaintext raw PED file (see \code{--recodeA}
in PLINK) or PED-like file for testing purposes.
}

