% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMatrix.R
\docType{methods}
\name{apply,dMatrix-method}
\alias{apply,dMatrix-method}
\title{Apply function for \code{\linkS4class{rDMatrix}} or
\code{\linkS4class{cDMatrix}} objects.}
\usage{
\S4method{apply}{dMatrix}(X, MARGIN, FUN, chunkSize = 1000, verbose = TRUE,
  ...)
}
\arguments{
\item{X}{Either an \code{\linkS4class{rDMatrix}} or a
\code{\linkS4class{cDMatrix}} object.}

\item{MARGIN}{Use 1 to obtain row summaries or 2 to obtain column summaries.}

\item{chunkSize}{The number of columns or rows that are processed at a time
(see Details).}
}
\value{
Returns a \code{matrix} or a \code{list} with results from FUN.
}
\description{
This function brings chunks of data (of size \code{chunkSize}) from the
distributed array into RAM as \code{matrix} objects and calls \code{apply} of
the base package to obtain the summaries for the chunk. Results from all the
chunks are collected and returned.
}

