% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{as.BGData.BEDMatrix}
\alias{as.BGData.BEDMatrix}
\title{Convert a \code{BEDMatrix} object into a \code{\link[=BGData-class]{BGData}}
object.}
\usage{
\method{as.BGData}{BEDMatrix}(x, alternatePhenotypeFile = NULL, ...)
}
\arguments{
\item{x}{A \code{BEDMatrix} object.}

\item{alternatePhenotypeFile}{Path to an alternate phenotype file (see PLINK 
documentation).}

\item{...}{Additional arguments to the \code{BEDMatrix} constructor
(currently unused).}
}
\value{
A \code{\link[=BGData-class]{BGData}} object.
}
\description{
Convert a \code{BEDMatrix} object into a \code{\link[=BGData-class]{BGData}}
object.
}

