% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{as.BGData.BEDMatrix}
\alias{as.BGData.BEDMatrix}
\title{Converts a \code{BEDMatrix} object to a \code{\link[=BGData-class]{BGData}}
object.}
\usage{
\method{as.BGData}{BEDMatrix}(x, alternatePhenotypeFile = NULL, ...)
}
\arguments{
\item{x}{A \code{BEDMatrix} object.}

\item{alternatePhenotypeFile}{Path to an
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#pheno}{alternate
 phenotype file}.}

\item{...}{Additional arguments to the \code{BEDMatrix} constructor
(currently unused).}
}
\value{
A \code{\link[=BGData-class]{BGData}} object.
}
\description{
If a FAM file (which corresponds to the first six columns of a PED file) of
the same name and in the same directory as the BED file exists, the
\code{@pheno} slot will be populated with the data stored in that file.
Otherwise a stub that only contains an \code{IID} column populated with the
rownames of \code{@geno} will be generated.
}
\details{
PED only allows for a single phenotype. If more phenotypes are required it is
possible to store them in an
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#pheno}{alternate
phenotype file}. The path to such a file can be provided with
\code{alternatePhenotypeFile} and will be merged with the data in the
\code{@pheno} slot.
}

