% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{as.BGData.BEDMatrix}
\alias{as.BGData.BEDMatrix}
\title{Converts Other Objects to BGData Objects.}
\usage{
\method{as.BGData}{BEDMatrix}(x, alternatePhenotypeFile = NULL, ...)
}
\arguments{
\item{x}{A \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix} object.}

\item{alternatePhenotypeFile}{Path to an \href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#pheno}{alternate phenotypefile}.}

\item{...}{Additional arguments to the \code{\link[utils:read.table]{utils::read.table()}} or
\code{\link[data.table:fread]{data.table::fread()}} call (if data.table package is installed) call to
parse the alternate pheno file.}
}
\value{
A \linkS4class{BGData} object.
}
\description{
If a FAM file (which corresponds to the first six columns of a PED file) of
the same name and in the same directory as the BED file exists, the
\code{@pheno} slot will be populated with the data stored in that file.
Otherwise a stub that only contains an \code{IID} column populated with the
rownames of \code{@geno} will be generated.
}
\details{
PED only allows for a single phenotype. If more phenotypes are required it
is possible to store them in an \href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#pheno}{alternate phenotypefile}. The path
to such a file can be provided with \code{alternatePhenotypeFile} and will be
merged with the data in the \code{@pheno} slot.
}
\seealso{
\code{\link[=readPED]{readPED()}} to convert text files to \linkS4class{BGData} objects.
}
