% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{as.BGData.BEDMatrix}
\alias{as.BGData.BEDMatrix}
\title{Convert a BEDMatrix object into a BGData object.}
\usage{
\method{as.BGData}{BEDMatrix}(x, ...)
}
\arguments{
\item{x}{A BEDMatrix object.}

\item{...}{Additional arguments to the BEDMatrix constructor (currently unused).}
}
\value{
A BGData object.
}
\description{
Convert a BEDMatrix object into a BGData object.
}

