% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{as.BGData}
\alias{as.BGData}
\title{S3 generic to convert other objects into BGData objects.}
\usage{
as.BGData(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments (see method).}
}
\value{
A BGData object.
}
\description{
S3 generic to convert other objects into BGData objects.
}

