% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{as.BGData}
\alias{as.BGData}
\title{S3 Generic to Convert Other Objects into BGData Objects.}
\usage{
as.BGData(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments (see method).}
}
\value{
A BGData object.
}
\description{
S3 Generic to Convert Other Objects into BGData Objects.
}
