% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BGData.R
\docType{class}
\name{BGData-class}
\alias{BGData}
\alias{BGData-class}
\title{An S4 class to represent GWAS data.}
\description{
An S4 class to represent GWAS data.
}
\section{Slots}{

\describe{
\item{\code{geno}}{A \code{geno} object (\code{LinkedMatrix}, \code{ff_matrix}, or
\code{\link{matrix}}) that contains genotypes.}

\item{\code{pheno}}{A \code{\link{data.frame}} that contains phenotypes.}

\item{\code{map}}{A \code{\link{data.frame}} that contains a genetic map.}
}}

