% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BGData-package}
\alias{BGData-package}
\title{A Suite of Packages for Analysis of Big Genomic Data.}
\description{
Modern genomic datasets are big (large \emph{n}), high-dimensional (large \emph{p}),
and multi-layered. The challenges that need to be addressed are memory
requirements and computational demands. Our goal is to develop software that
will enable researchers to carry out analyses with big genomic data within
the R environment.
}
\details{
We have identified several approaches to tackle those challanges within R:
\itemize{
\item Memory mapping: The data is stored in on the hard drive and users can read
in smaller chunks when they are needed.
\item Linked arrays: For very large datasets a single memory-mapped array may
not be enough or convenient. A linked array is an array whose content is
distributed over multiple memory-mapped nodes.
\item Multiple dispatch: Methods are presented to users so that they can treat
these arrays pretty much as if they were RAM arrays.
\item Multi-level parallelism: Exploit multi-core and multi-node computing.
\item Inputs: Users can create these arrays from standard formats (e.g., PLINK
.bed).
}

The BGData package is an umbrella package that comprises several packages:
\link[BEDMatrix:BEDMatrix-package]{BEDMatrix},
\link[LinkedMatrix:LinkedMatrix-package]{LinkedMatrix}, and
\link[symDMatrix:symDMatrix-package]{symDMatrix}. It features scalable and
efficient computational methods for large genomic datasets such as
genome-wide association studies (GWAS) or genomic relationship matrices (G
matrix). It also contains a data structure called \code{BGData} that holds
genotypes in the \code{@geno} slot, phenotypes in the \code{@pheno} slot, and
additional information in the \code{@map} slot.

@seealso \link[BEDMatrix:BEDMatrix-package]{BEDMatrix::BEDMatrix-package},
\link[LinkedMatrix:LinkedMatrix-package]{LinkedMatrix::LinkedMatrix-package}, and \link[symDMatrix:symDMatrix-package]{symDMatrix::symDMatrix-package}
for an introduction to the respective packages.
}
