% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunkedApply}
\alias{chunkedApply}
\title{Reads chunks of data from a memory-mapped file into memory and applies a
function on each row or column of a matrix in parallel.}
\usage{
chunkedApply(X, MARGIN, FUN, bufferSize, nTasks = parallel::detectCores(),
  mc.cores = parallel::detectCores(), verbose = FALSE, ...)
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{MARGIN}{The subscripts which the function will be applied over. 1
indicates rows, 2 indicates columns.}

\item{FUN}{The function to be applied.}

\item{bufferSize}{The number of rows or columns of \code{X} that are brought
into memory for processing.}

\item{nTasks}{The number of submatrices of each buffered subset of \code{X}
to be processed in parallel.}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}

\item{verbose}{Whether to print additional information.}

\item{...}{Additional arguments to be passed to \code{parallelApply}.}
}
\description{
This function is only useful for memory-mapped files. For data that is
already in memory, use \code{\link{parallelApply}} directly.
}

