% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunkedApply}
\alias{chunkedApply}
\title{Reads chunks of data into memory and applies a function on each row or
column of a matrix.}
\usage{
chunkedApply(X, MARGIN, FUN, bufferSize, verbose = FALSE, ...)
}
\arguments{
\item{X}{A matrix.}

\item{MARGIN}{The subscripts which the function will be applied over. 1
indicates rows, 2 indicates columns.}

\item{FUN}{The function to be applied.}

\item{bufferSize}{The number of rows or columns of \code{X} that are brought
into memory for processing.}

\item{verbose}{Whether to print additional information.}

\item{...}{Additional arguments to be passed to \code{parallelApply}.}
}
\description{
\code{chunkedApply} uses \code{parallelApply} internally, so \code{nTasks}
and \code{mc.cores} can be passed as \code{...}.
}

