% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMatrix.R
\name{chunks}
\alias{chunks}
\title{Provides information about how data is distributed into binary files.}
\usage{
chunks(x)
}
\arguments{
\item{x}{Either an rDMatrix or a cDMatrix object}
}
\value{
A matrix with information per chunk in rows.
}
\description{
Row-distributed (rDMatrix) and column-distributed matrices (cDMatrix) have
the content of the array mapped to possibly multiple binary files. Each chunk
is an ff_matrix object. chunks() gives, for each chunk, the row or column
indexes at which each chunk start and ends.
}

