% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossprod.parallel}
\alias{crossprod.parallel}
\title{Computes crossprod (x'y or x'x) in parallel.}
\usage{
crossprod.parallel(x, y = NULL, nChunks = parallel::detectCores(),
  mc.cores = parallel::detectCores())
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{y}{vector or matrix-like object. NULL by default.}

\item{nChunks}{The number of chunks used when X and y are partitioned.}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\value{
x'y' or x'x depending on whether y is provided.
}
\description{
Computes crossprod (x'y or x'x) in parallel.
}

