% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossprod.parallel}
\alias{crossprod.parallel}
\title{Computes crossprod (x'y or x'x) in parallel.}
\usage{
crossprod.parallel(x, y = NULL, nTasks = nCores,
  nCores = parallel::detectCores())
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{y}{vector or matrix-like object. NULL by default.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\value{
x'y' or x'x depending on whether y is provided.
}
\description{
Computes crossprod (x'y or x'x) in parallel.
}

