% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossprod.parallel}
\alias{crossprod.parallel}
\title{Computes crossprod (x'y or x'x) in Parallel.}
\usage{
crossprod.parallel(x, y = NULL, nTasks = nCores,
  nCores = parallel::detectCores())
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a \linkS4class{BGData}
object.}

\item{y}{vector or matrix-like object. NULL by default.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores. Defaults to \code{nCores}.}

\item{nCores}{The number of cores (passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}).
Defaults to the number of cores as detected by \code{\link[parallel:detectCores]{parallel::detectCores()}}.}
}
\value{
x'y or x'x depending on whether y is provided.
}
\description{
Computes crossprod (x'y or x'x) in Parallel.
}
