% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{crossprod.parallel}
\alias{crossprod.parallel}
\title{Computes crossprod (x'y or x'x) in parallel.}
\usage{
crossprod.parallel(x, y = NULL, nChunks = detectCores(),
  mc.cores = detectCores())
}
\arguments{
\item{x}{matrix, ff_matrix, RowLinkedMatrix or ColumnLinkedMatrix.}

\item{y}{vector, matrix, ff_matrix, RowLinkedMatrix or ColumnLinkedMatrix. NULL by default.}

\item{nChunks}{The number of chunks used when X and y are partitioned.}

\item{mc.cores}{The number of cores (passed to \code{\link[parallel]{mclapply}}).}
}
\value{
x'y' or x'x depending on whether y is provided.
}
\description{
Computes crossprod (x'y or x'x) in parallel.
}

