% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRelated.R
\name{findRelated}
\alias{findRelated}
\title{Find related individuals.}
\usage{
findRelated(x, ...)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object with dimnames.}

\item{cutoff}{The cutoff between 0 and 1 for related individuals to be included
in the output. Defaults to 0.03.}

\item{verbose}{Whether progress updates will be posted. Defaults to \code{FALSE}.}
}
\value{
A vector of names or indices of related individuals.
}
\description{
Find related individuals.
}
\examples{
# Load example data
bg <- BGData:::loadExample()

G <- getG(bg@geno)
findRelated(G)
}
