\name{FWD}
\alias{FWD}
\title{Performs Forward Regressions}
\description{
    TODO
}
\usage{
FWD(y, X, df = 20, tol = 1e-7, verbose = TRUE, maxIter = 1000, centerImpute =
  TRUE)
}
\arguments{
    \item{y}{
    }
    \item{X}{
    }
    \item{df}{
    }
    \item{tol}{
    }
    \item{verbose}{
    }
    \item{maxIter}{
    }
    \item{centerImpute}{
    }
}
\value{
    TODO
}
