% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getG}
\alias{getG}
\title{Computes a genomic relationship matrix G=xx'.}
\usage{
getG(x, nChunks = ceiling(ncol(x)/10000), scaleCol = TRUE,
  centerCol = TRUE, scaleG = TRUE, verbose = TRUE, i = seq_len(nrow(x)),
  j = seq_len(ncol(x)), i2 = NULL, minVar = 1e-05, nTasks = mc.cores,
  scales = NULL, centers = NULL, saveG = FALSE, saveType = "RData",
  saveName = "Gij", mc.cores = parallel::detectCores())
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{nChunks}{The number of columns that are processed at a time.}

\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
xx'.}

\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
xx'.}

\item{scaleG}{TRUE/FALSE whether xx' must be scaled.}

\item{verbose}{If TRUE more messages are printed.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{i2}{(integer, boolean or character) Indicates which rows should be used
to divide matrix into blocks.}

\item{minVar}{Columns with variance lower than this value will not be used in
the computation (only if \code{scaleCol} is set).}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{mc.cores} cores.}

\item{scales}{Precomputed scales if i2 is used.}

\item{centers}{Precomputed centers if i2 is used.}

\item{saveG}{Whether to save genomic relationship matrix into file.}

\item{saveType}{File format to save genomic relationship matrix in. Either
\code{RData} or \code{ff}.}

\item{saveName}{Name without extension to save genomic relationship matrix
with.}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\value{
A positive semi-definite symmetric numeric matrix.
}
\description{
Offers options for centering and scaling the columns of x before computing
xx'. If \code{centerCol=FALSE}, \code{scaleCol=FALSE} and
\code{scaleG=FALSE}, \code{getG} produces the same outcome than
\code{tcrossprod}.
}

