% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{getG}
\alias{getG}
\title{Computes a genomic relationship matrix G=XX'.}
\usage{
getG(x, nChunks = 3, scaleCol = TRUE, scaleG = TRUE, verbose = TRUE,
  i = 1:nrow(x), j = 1:ncol(x), minVar = 1e-05)
}
\arguments{
\item{x}{matrix, ff_matrix, rmmMatrix or cmmMatrix}

\item{nChunks}{The number of columns that are processed at a time.}
\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
XX'.}
\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
XX'.}

\item{scaleG}{TRUE/FALSE whether columns must be scaled before computing XX'.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}
}
\value{
Genomic relationship matrix
}
\description{
Offers options for centering and scaling G=WW' where \code{W <- scale(X,
center=centerCol, scale=scaleCol)} and of scaling the final output so that
the average diagonal value is equal to one (\code{scaleG=TRUE}). If
\code{scaleCol=centerCol} and \code{scaleG=FALSE} it behaves as
\code{tcrossprod(X)}.
\code{getG(x,scaleCol=FALSE,centerCol=FALSE,scaleG=FALSE)} should be equal than \code{tcrossprod(x)}.
}

