% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getG}
\alias{getG}
\title{Computes a genomic relationship matrix G=xx'.}
\usage{
getG(x, scaleCol = TRUE, scales = NULL, centerCol = TRUE,
  centers = NULL, scaleG = TRUE, minVar = 1e-05, saveG = FALSE,
  saveType = "RData", folderOut = paste0("G_", randomString()),
  saveName = "Gij", i = seq_len(nrow(x)), j = seq_len(ncol(x)),
  i2 = NULL, bufferSize = 5000, nBuffers = NULL, nTasks = nCores,
  nCores = parallel::detectCores(), verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
xx'.}

\item{scales}{Precomputed scales if i2 is used.}

\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
xx'.}

\item{centers}{Precomputed centers if i2 is used.}

\item{scaleG}{TRUE/FALSE whether xx' must be scaled.}

\item{minVar}{Columns with variance lower than this value will not be used in
the computation (only if \code{scaleCol} is set).}

\item{saveG}{Whether to save genomic relationship matrix into file.}

\item{saveType}{File format to save genomic relationship matrix in. Either
\code{RData} or \code{ff}.}

\item{folderOut}{The path to the folder where to save the output generated
by \code{saveType}. Defaults to a random string prefixed with "G_".}

\item{saveName}{Name without extension to save genomic relationship matrix
with.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{i2}{(integer, boolean or character) Indicates which rows should be used
to divide matrix into blocks.}

\item{bufferSize}{The number of columns of \code{x} that are brought into
RAM for processing. Overwrites \code{nBuffers}. If both parameters are
\code{NULL}, all columns of \code{x} are used. Defaults to 5000.}

\item{nBuffers}{The number of partitions of the columns of \code{x} that are
brought into RAM for processing. Is overwritten by \code{bufferSize}. If
both parameters are \code{NULL}, all columns of \code{x} are used.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores. Defaults to \code{nCores}.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}). Defaults to the number of cores as
detected by \code{\link[parallel]{detectCores}}).}

\item{verbose}{Whether progress updates will be posted. Defaults to
\code{TRUE}.}
}
\value{
A positive semi-definite symmetric numeric matrix.
}
\description{
Offers options for centering and scaling the columns of x before computing
xx'. If \code{centerCol=FALSE}, \code{scaleCol=FALSE} and
\code{scaleG=FALSE}, \code{getG} produces the same outcome than
\code{tcrossprod}.
}

