% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getG.symDMatrix}
\alias{getG.symDMatrix}
\title{Computes a genomic relationship matrix G=xx' without ever loading G in RAM by
creating a \code{\link[=symDMatrix-class]{symDMatrix}}.}
\usage{
getG.symDMatrix(X, nChunks = 5, chunkSize = NULL, centers = NULL,
  scales = NULL, centerCol = T, scaleCol = T, scaleG = T,
  nChunks2 = 1, folder = randomString(), vmode = "double",
  verbose = TRUE, saveRData = TRUE, mc.cores = parallel::detectCores())
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{nChunks}{The number of columns that are processed at a time.}

\item{chunkSize}{The number of columns that are processed at a time.}

\item{centers}{Precomputed centers.}

\item{scales}{Precomputed scales.}

\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
xx'.}

\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
xx'.}

\item{scaleG}{TRUE/FALSE whether xx' must be scaled.}

\item{nChunks2}{The number of chunks that each chunk is split into for
processing in parallel.}

\item{folder}{Folder in which to save the
\code{\link[=symDMatrix-class]{symDMatrix}}.}

\item{vmode}{vmode of \code{ff} objects.}

\item{verbose}{If TRUE more messages are printed.}

\item{saveRData}{Whether to save an RData file to easily reload
\code{\link[=symDMatrix-class]{symDMatrix}}}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\value{
A positive semi-definite symmetric numeric matrix.
}
\description{
Offers options for centering and scaling the columns of x before computing
xx'.
}

