% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getG.symDMatrix}
\alias{getG.symDMatrix}
\title{Computes a genomic relationship matrix G=xx' without ever loading G in RAM by
creating a \code{\link[=symDMatrix-class]{symDMatrix}}.}
\usage{
getG.symDMatrix(X, nBlocks = 5, blockSize = NULL, centers = NULL,
  scales = NULL, centerCol = TRUE, scaleCol = TRUE, scaleG = TRUE,
  folderOut = paste0("symDMatrix_", randomString()), vmode = "double",
  saveRData = TRUE, i = seq_len(nrow(X)), j = seq_len(ncol(X)),
  nTasks = nCores, nCores = parallel::detectCores(), verbose = TRUE)
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{nBlocks}{The number of blocks.}

\item{blockSize}{The number of columns of a block (if NULL inferred from block).}

\item{centers}{Precomputed centers.}

\item{scales}{Precomputed scales.}

\item{centerCol}{TRUE/FALSE whether columns must be centered before computing
xx'.}

\item{scaleCol}{TRUE/FALSE whether columns must be scaled before computing
xx'.}

\item{scaleG}{TRUE/FALSE whether xx' must be scaled.}

\item{folderOut}{The path to the folder where to save the
\code{\link[=symDMatrix-class]{symDMatrix}} object. Defaults to a random
string prefixed with "symDMatrix_".}

\item{vmode}{vmode of \code{ff} objects.}

\item{saveRData}{Whether to save an RData file to easily reload
\code{\link[=symDMatrix-class]{symDMatrix}}}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores. Defaults to \code{nCores}.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}). Defaults to the number of cores as
detected by \code{\link[parallel]{detectCores}}).}

\item{verbose}{Whether progress updates will be posted. Defaults to
\code{TRUE}.}
}
\value{
A positive semi-definite symmetric numeric matrix.
}
\description{
Offers options for centering and scaling the columns of x before computing
xx'.
}

