% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GWAS}
\alias{GWAS}
\title{Performs single marker regressions using a
\code{\link[=BGData-class]{BGData}} object.}
\usage{
GWAS(formula, data, method, i = seq_len(nrow(data@geno)),
  j = seq_len(ncol(data@geno)), verbose = FALSE, chunkSize = 5000,
  nTasks = nCores, nCores = parallel::detectCores(), ...)
}
\arguments{
\item{formula}{A formula (e.g. weight~sex+age) with the response on the
left-hand side and predictors (all the covariates except the markers) on
the right-hand side. The variables included in the formula must be in the
\code{@pheno} object of the \code{\link[=BGData-class]{BGData}}.}

\item{data}{A \code{\link[=BGData-class]{BGData}} object.}

\item{method}{The regression method to be used. Currently, the following
methods are implemented: \code{\link{lm}}, \code{\link{lm.fit}},
\code{\link{lsfit}}, \code{\link{glm}} and \code{\link[lme4]{lmer}}.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{verbose}{If TRUE more messages are printed.}

\item{chunkSize}{Represents the number of columns of \code{@geno} that are
brought into RAM for processing (5000 by default).}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}

\item{...}{Additional arguments for chunkedApply and regression method.}
}
\value{
Returns a matrix with estimates, SE, p-value, etc.
}
\description{
Implements single marker regressions. The regression model includes all the
covariates specified in the right-hand-side of the \code{formula} plus one
column of \code{@geno}, one column at a time. The data from the association
tests is obtained from a \code{\link[=BGData-class]{BGData}} object.
}

