% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GWAS}
\alias{GWAS}
\title{Performs single marker regressions using a \code{\linkS4class{BGData}} 
object.}
\usage{
GWAS(formula, data, method, plot = FALSE, verbose = FALSE,
  min.pValue = 1e-10, chunkSize = 5000, ...)
}
\arguments{
\item{formula}{A formula (e.g. weight~sex+age) with the response on the 
left-hand side and predictors (all the covariates except the markers) on 
the right-hand side. The variables included in the formula must be in the 
\code{@pheno} object of the \code{\linkS4class{BGData}}.}

\item{data}{A \code{\linkS4class{BGData}} object.}

\item{method}{The regression method to be used. Currently, the following 
methods are implemented: \code{\link{lm}}, \code{\link{lm.fit}}, 
\code{\link{lsfit}}, \code{\link{glm}} and \code{\link[lme4]{lmer}}.}

\item{plot}{If TRUE a Manhattan plot is produced and filled with points as 
the association tests are run.}

\item{verbose}{If TRUE more messages are printed.}

\item{min.pValue}{Numeric, the minimum p-value expected, used to determine 
the limits of the vertical axis of the Manhattan plot.}

\item{chunkSize}{Represents the number of columns of \code{@geno} that are 
brought into RAM for processing (5000 by default).}

\item{...}{Optional arguments for chunkedApply and regression method.}
}
\value{
Returns a matrix with estimates, SE, p-value, etc.
}
\description{
Implements single marker regressions. The regression model includes all the
covariates specified in the right-hand-side of the \code{formula} plus one
column of \code{@geno}, one column at a time. The data from the association
tests is obtained from a \code{\linkS4class{BGData}} object.
}

