% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\docType{methods}
\name{initialize,BGData-method}
\alias{initialize,BGData-method}
\title{Creates a new \code{\linkS4class{BGData}} instance.}
\usage{
\S4method{initialize}{BGData}(.Object, geno, pheno, map)
}
\arguments{
\item{.Object}{The \code{ColumnLinkedMatrix} instance to be initialized.}

\item{geno}{A \code{geno} object (\code{\linkS4class{LinkedMatrix}},
  \code{BEDMatrix}, \code{\linkS4class{big.matrix}}, \code{ff_matrix}, or
\code{\link{matrix}}) that contains genotypes.}

\item{pheno}{A \code{\link{data.frame}} that contains phenotypes. A stub that
only contains an \code{IID} column populated with the rownames of
\code{@geno} will be generated if missing.}

\item{map}{A \code{\link{data.frame}} that contains a genetic map. A stub
that only contains a \code{mrk} column populated with the colnames of
\code{@geno} will be generated if missing.}
}
\description{
Creates a new \code{\linkS4class{BGData}} instance.
}

