% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BGData.R
\name{load.BGData}
\alias{load.BGData}
\title{Loads BGData objects.}
\usage{
load.BGData(file, envir = parent.frame(), verbose = TRUE)
}
\arguments{
\item{file}{The name of the .RData file to be loaded.}

\item{envir}{The environment where to load the data.}

\item{verbose}{TRUE/FALSE}
}
\description{
Loads BGData objects.
}

