% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{load.BGData}
\alias{load.BGData}
\title{Loads BGData objects.}
\usage{
load.BGData(file, envir = parent.frame())
}
\arguments{
\item{file}{The name of the .RData file to be loaded.}

\item{envir}{The environment where to load the data.}
}
\description{
If the value in the \code{@geno} slot is of class 
\code{\linkS4class{LinkedMatrix}} and the nodes contain \code{ff} objects 
\code{load.BGData} will attempt to open them. If the value is of class 
\code{\linkS4class{big.matrix}}, it will attempt to attach the matrix.
}

