% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{readPED.big.matrix}
\alias{readPED.big.matrix}
\title{Creates a \code{\linkS4class{BGData}} object from a plaintext PED-like file.}
\usage{
readPED.big.matrix(fileIn, header, dataType, n = NULL, p = NULL,
  na.strings = "NA", nColSkip = 6, idCol = c(1, 2), verbose = FALSE,
  folderOut = paste("BGData_", sub("\\\\.[[:alnum:]]+$", "",
  basename(fileIn)), sep = ""))
}
\arguments{
\item{fileIn}{The path to the plaintext file.}

\item{header}{If TRUE, the file contains a header.}

\item{dataType}{The coding of genotypes. Use \code{character()} for A/C/G/T 
or \code{integer()} for numeric coding.}

\item{n}{The number of individuals.}

\item{p}{The number of markers.}

\item{na.strings}{The character string used in the plaintext file to denote 
missing value.}

\item{nColSkip}{The number of columns to be skipped to reach the genotype 
information in the file.}

\item{idCol}{The index of the ID column. If more than one index is given, 
both columns will be concatenated with "_".}

\item{verbose}{If TRUE, progress updates will be posted.}

\item{folderOut}{The path to the folder where to save the binary files.}
}
\value{
Returns a \code{\linkS4class{BGData}} object.
}
\description{
Genotypes are stored in a filebacked \code{\linkS4class{big.matrix}} object. 
\code{readPED.big.matrix} creates a folder (see \code{folderOut}) that
contains the binary flat file (named \code{BGData.bin}), a descriptor file
(named \code{BGData.desc}), and an external representation of the
\code{\linkS4class{BGData}} object in \code{BGData.RData}. A 
\code{\linkS4class{BGData}} object can be reloaded using \code{load.BGData} 
(the regular \code{load} function will not work unless the
\code{\linkS4class{big.matrix}} instance is manually attached using
\code{\link[bigmemory]{attach.big.matrix}}).
}
\seealso{
\code{\linkS4class{BGData}}
}

