% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGData.R
\name{readPED.big.matrix}
\alias{readPED.big.matrix}
\title{Creates a BGData Object from a Raw PED File (Generated with --recodeA in
PLINK) or a PED-Like File and Stores the Genotypes in a big.matrix.}
\usage{
readPED.big.matrix(fileIn, header = TRUE, dataType = integer(), n = NULL,
  p = NULL, sep = "", na.strings = "NA", nColSkip = 6, idCol = c(1,
  2), folderOut = paste0("BGData_", sub("\\\\.[[:alnum:]]+$", "",
  basename(fileIn))), outputType = "char", verbose = FALSE)
}
\arguments{
\item{fileIn}{The path to the plaintext file.}

\item{header}{Whether \code{fileIn} contains a header. Defaults to \code{TRUE}.}

\item{dataType}{The coding type of genotypes in \code{fileIn}. Use \code{integer()} or
\code{double()} for numeric coding. Alpha-numeric coding is currently not
supported for \code{\link[=readPED]{readPED()}}: use the \code{--recodeA} option of PLINK to convert
the PED file into a raw file. Defaults to \code{integer()}.}

\item{n}{The number of individuals.}

\item{p}{The number of markers.}

\item{sep}{The field separator character. Values on each line of the file
are separated by this character. If \code{sep = ""} (the default for
\code{\link[=readPED.big.matrix]{readPED.big.matrix()}} the separator is "white space", that is one or more
spaces, tabs, newlines or carriage returns.}

\item{na.strings}{The character string used in the plaintext file to denote
missing value.}

\item{nColSkip}{The number of columns to be skipped to reach the genotype
information in the file.}

\item{idCol}{The index of the ID column. If more than one index is given,
both columns will be concatenated with "_".}

\item{folderOut}{The path to the folder where to save the binary files.}

\item{outputType}{The type of the \link[bigmemory:big.matrix-class]{bigmemory::big.matrix}.}

\item{verbose}{Whether progress updates will be posted. Defaults to \code{FALSE}.}
}
\value{
Returns a \linkS4class{BGData} object.
}
\description{
\code{\link[=readPED.matrix]{readPED.matrix()}} assumes that the plaintext file (\code{fileIn}) contains
records of individuals in rows, and phenotypes, covariates and markers in
columns. The columns included in columns \code{seq_len(nColSkip)} are used to
populate the slot \code{@pheno} of a \linkS4class{BGData} object, and the remaining
columns are used to fill the slot \code{@geno}. If the first row contains a
header (\code{header=TRUE}), data in this row is used to determine variables
names for \code{@pheno} and marker names for \code{@map} and \code{@geno}.
}
\details{
Genotypes are stored in a filebacked \link[bigmemory:big.matrix-class]{bigmemory::big.matrix} object.
\code{\link[=readPED.big.matrix]{readPED.big.matrix()}} creates a folder (see \code{folderOut}) that contains the
binary flat file (named \code{BGData.bin}), a descriptor file (named
\code{BGData.desc}), and an external representation of the \linkS4class{BGData} object
in \code{BGData.RData}. A \linkS4class{BGData} object can be reloaded using
\code{\link[=load.BGData]{load.BGData()}} (the regular \code{\link[base:load]{base::load()}} function will not work unless
the \link[bigmemory:big.matrix-class]{bigmemory::big.matrix} instance is manually attached using
\code{\link[bigmemory:attach.big.matrix]{bigmemory::attach.big.matrix()}}).
}
\seealso{
\linkS4class{BGData}
}
