% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BGData.R
\name{readPED.matrix}
\alias{readPED.matrix}
\title{Creates a \code{\linkS4class{BGData}} object from a plaintext PED-like file.}
\usage{
readPED.matrix(fileIn, header, dataType, n = NULL, p = NULL,
  na.strings = "NA", nColSkip = 6, idCol = 2, verbose = FALSE)
}
\arguments{
\item{fileIn}{The path to the plaintext file.}

\item{header}{If TRUE, the file contains a header.}

\item{dataType}{The coding of genotypes. Use 'character' for A/C/G/T or
'integer' for numeric coding.}

\item{n}{The number of individuals.}

\item{p}{The number of markers.}

\item{na.strings}{The character string used in the plaintext file to denote
missing value.}

\item{nColSkip}{The number of columns to be skipped to reach the genotype
information in the file.}

\item{idCol}{The index of the ID column.}

\item{verbose}{If TRUE, progress updates will be posted.}
}
\value{
Returns a \code{\linkS4class{BGData}} object.
}
\description{
\code{readPED.matrix} assumes that the plaintext file (\code{fileIn})
contains records of individuals in rows, and phenotypes, covariates and
markers in columns. The columns included in columns \code{1:nColSkip} are
used to populate the slot \code{\code{@pheno}} of a
\code{\linkS4class{BGData}} object, and the remaining columns are used to
fill the slot \code{\code{@geno}}. If the first row contains a header
(\code{header=TRUE}), data in this row is used to determine variables names
for \code{@pheno} and marker names for \code{@map} and \code{@geno}.
}
\seealso{
\code{\linkS4class{BGData}}
}

