% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BGData.R
\name{readPED}
\alias{readPED}
\title{Creates a memory-mapped \code{\linkS4class{BGData}} object from a plaintext
raw PED file (generated with \code{--recodeA} in PLINK) or PED-like file.}
\usage{
readPED(fileIn, header, dataType, n = NULL, p = NULL, na.strings = "NA",
  nColSkip = 6, idCol = 2, returnData = TRUE, verbose = FALSE,
  nChunks = NULL, distributed.by = "rows", folderOut = paste("BGData_",
  sub("\\\\.[[:alnum:]]+$", "", basename(fileIn)), sep = ""), dimorder = if
  (distributed.by == "rows") 2:1 else 1:2)
}
\arguments{
\item{fileIn}{The path to the plaintext file.}

\item{header}{If TRUE, the file contains a header.}

\item{dataType}{The coding of genotypes. Use \code{integer()} or
\code{double()} for numeric coding. Character coding is currently not
supported: use the \code{--recodeA} option of PLINK to convert the PED file
into a raw file.}

\item{n}{The number of individuals.}

\item{p}{The number of markers.}

\item{na.strings}{The character string used in the plaintext file to denote
missing value.}

\item{nColSkip}{The number of columns to be skipped to reach the genotype
information in the file.}

\item{idCol}{The index of the ID column.}

\item{returnData}{If TRUE, the function returns a \code{\linkS4class{BGData}}
object.}

\item{verbose}{If TRUE, progress updates will be posted.}

\item{nChunks}{The number of chunks to create.}

\item{distributed.by}{If columns a column-distributed matrix
(\code{\linkS4class{ColumnLinkedMatrix}}) is created, if rows a row-distributed
matrix (\code{\linkS4class{RowLinkedMatrix}}).}

\item{folderOut}{The path to the folder where to save the binary files.}

\item{dimorder}{The physical layout of the chunks.}
}
\value{
If \code{returnData} is TRUE, a \code{\linkS4class{BGData}} object is
  returned.
}
\description{
\code{readPED} assumes that the plaintext file (\code{fileIn}) contains
records of individuals in rows, and phenotypes, covariates and markers in
columns. The columns included in columns \code{1:nColSkip} are used to
populate the slot \code{@pheno} of a \code{\linkS4class{BGData}} object, and
the remaining columns are used to fill the slot \code{@geno}. If the first
row contains a header (\code{header=TRUE}), data in this row is used to
determine variables names for \code{@pheno} and marker names for
\code{@map} and \code{@geno}. Genotypes are stored in a distributed matrix
(\code{LinkedMatrix}). By default a column-distributed
(\code{\linkS4class{ColumnLinkedMatrix}}) is used for \code{@geno}, but the user can
modify this using the \code{distributed.by} argument. The number of chunks is
either specified by the user (use \code{nChunks} when calling \code{readPED})
or determined internally so that each \code{ff_matrix} object has a number of
cells that is smaller than \code{.Machine$integer.max/1.2}. \code{readPED}
creates a folder (\code{folderOut}) that contains the binary flat files
(\code{geno_*.bin}) and the \code{\linkS4class{BGData}} object (typically
named \code{BGData.RData}. Optionally (if \code{returnData} is TRUE) it
returns the \code{\linkS4class{BGData}} object to the environment. The
filename of the \code{ff_matrix} objects are saved as relative names.
Therefore, to be able to access the content of the data included in
\code{@geno} the working directory must either be the folder where these
files are saved (\code{folderOut}) or the object must be loaded using
\code{load.BGData}.
}
\seealso{
\code{\linkS4class{BGData}}, \code{LinkedMatrix},
  \code{\linkS4class{ColumnLinkedMatrix}}, \code{\linkS4class{RowLinkedMatrix}},
  \code{\link[ff]{ff}}
}

