% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmmMatrix.R
\name{rowindexes}
\alias{rowindexes}
\title{Finds the position of a set of rows in an rmmMatrix object.}
\usage{
rowindexes(x, rows)
}
\description{
Finds the position of a set of rows in an rmmMatrix object.
}

