% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize}
\alias{summarize}
\title{Calculates frequencies of missing values and alleles.}
\usage{
summarize(X, i = seq_len(nrow(X)), j = seq_len(ncol(X)),
  bufferSize = NULL, nBuffers = NULL, nTasks = nCores, nCores = 1,
  verbose = FALSE)
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{bufferSize}{The number of columns of \code{X} that are brought into RAM
for processing. Overwrites \code{nBuffers}. If both parameters are
\code{NULL}, all elements in \code{j} are used.}

\item{nBuffers}{The number of partitions of the columns of \code{X} that are
brought into RAM for processing. Is overwritten by \code{bufferSize}. If
both parameters are \code{NULL}, all elements in \code{j} are used.}

\item{nTasks}{Represents the number of parallel tasks each buffer is split
into.}

\item{nCores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}

\item{verbose}{If TRUE more messages are printed.}
}
\description{
Calculates frequencies of missing values and alleles.
}

