% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize}
\alias{summarize}
\title{Generates Various Summary Statistics.}
\usage{
summarize(X, i = seq_len(nrow(X)), j = seq_len(ncol(X)),
  bufferSize = 5000, nTasks = nCores, nBuffers = NULL,
  nCores = parallel::detectCores(), verbose = FALSE)
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a \linkS4class{BGData}
object.}

\item{i}{Indicates which rows of \code{X} should be used. Can be integer,
boolean, or character. By default, all rows are used.}

\item{j}{Indicates which columns of \code{X} should be used. Can be integer,
boolean, or character. By default, all columns are used.}

\item{bufferSize}{The number of columns of \code{X} that are brought into RAM for
processing. Overwrites \code{nBuffers}. If both parameters are \code{NULL}, all
elements in \code{j} are used. Defaults to 5000.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{nCores} cores. Defaults to \code{nCores}.}

\item{nBuffers}{The number of partitions of the columns of \code{X} that are
brought into RAM for processing. Is overwritten by \code{bufferSize}. If both
parameters are \code{NULL}, all elements in \code{j} are used. Defaults to \code{NULL}.}

\item{nCores}{The number of cores (passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}).
Defaults to the number of cores as detected by \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{verbose}{Whether progress updates will be posted. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame} with three columns: \code{freq_na} for frequencies of
missing values, \code{allele_freq} for (minor) allele frequencies, and \code{sd} for
standard deviations.
}
\description{
Computes the frequency of missing values, the (minor) allele frequency, and
standard deviation of each column of \code{X}.
}
