% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize}
\alias{summarize}
\title{Calculates frequencies of missing values and alleles.}
\usage{
summarize(X, verbose = FALSE, bufferSize = 5000, i = seq_len(nrow(X)),
  j = seq_len(ncol(X)), nTasks = mc.cores,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{X}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{verbose}{If TRUE more messages are printed.}

\item{bufferSize}{Represents the number of columns of \code{@geno} that are
brought into RAM for processing (5000 by default).}

\item{i}{(integer, boolean or character) Indicates which rows should be used.
By default, all rows are used.}

\item{j}{(integer, boolean or character) Indicates which columns should be
used. By default, all columns are used.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{mc.cores} cores.}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\description{
Calculates frequencies of missing values and alleles.
}

