% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcrossprod.parallel}
\alias{tcrossprod.parallel}
\title{Computes tcrossprod (xy' or xx') in parallel.}
\usage{
tcrossprod.parallel(x, y = NULL, nChunks = parallel::detectCores(),
  mc.cores = parallel::detectCores())
}
\arguments{
\item{x}{matrix, ff_matrix, RowLinkedMatrix or ColumnLinkedMatrix.}

\item{y}{vector, matrix, ff_matrix, RowLinkedMatrix or ColumnLinkedMatrix. NULL by default.}

\item{nChunks}{The number of chunks used when X and y are partitioned.}

\item{mc.cores}{The number of cores (passed to \code{\link[parallel]{mclapply}}).}
}
\value{
xy' or xx' depending on whether y is provided.
}
\description{
Computes tcrossprod (xy' or xx') in parallel.
}

