% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcrossprod.parallel}
\alias{tcrossprod.parallel}
\title{Computes tcrossprod (xy' or xx') in parallel.}
\usage{
tcrossprod.parallel(x, y = NULL, nTasks = mc.cores,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{x}{A matrix-like object, typically \code{@geno} of a
\code{\link[=BGData-class]{BGData}} object.}

\item{y}{vector or matrix-like object. NULL by default.}

\item{nTasks}{The number of tasks the problem should be broken into to be
distributed among \code{mc.cores} cores.}

\item{mc.cores}{The number of cores (passed to
\code{\link[parallel]{mclapply}}).}
}
\value{
xy' or xx' depending on whether y is provided.
}
\description{
Computes tcrossprod (xy' or xx') in parallel.
}

